requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
pyst2==0.5.1
APScheduler==3.10.4
PyMySQL==1.1.0
cryptography==41.0.5
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
six==1.16.0
tzlocal==5.0.1
pytz==2023.3
bcrypt==4.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
flask_cors
pyst2
